requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post1
fastapi==0.95.1
fonttools==4.39.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
orjson==3.8.10
packaging==23.1
passlib==1.7.4
peewee==3.16.0
Pillow==9.5.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pydicom==2.3.1
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
